soybean_pred/requirements.txt
numpy==1.16.2
Flask==1.0.3
Pillow==6.2.0
google-api-core==1.23.0
